ctutils

Trait BytesCtEq

pub trait BytesCtEq<Rhs: ?Sized = Self>: Sealed {
    // Required method
    fn bytes_ct_eq(&self, other: &Rhs) -> Choice;

    // Provided method
    fn bytes_ct_ne(&self, other: &Rhs) -> Choice { ... }
}

CtEq-like trait impl’d for [u8] and [u8; N] providing optimized implementations which perform than the generic impl of CtEq for [T; N] where T = u8.

Ideally we would use specialization to provide more specific impls of these traits for these types, but it’s unstable and unlikely to be stabilized soon.

Required Methods

fn bytes_ct_eq(&self, other: &Rhs) -> Choice

Determine if self is equal to the provided value.

Provided Methods

fn bytes_ct_ne(&self, other: &Rhs) -> Choice

Determine if self is NOT equal to the provided value.