ctutils

Trait BytesCtSelect

pub trait BytesCtSelect: Sized + Sealed {
    // Required methods
    fn bytes_ct_assign(&mut self, other: &Self, choice: Choice);
    fn bytes_ct_select(&self, other: &Self, choice: Choice) -> Self;
}

CtSelect-like trait impl’d for [u8] and [u8; N] providing optimized implementations which perform than the generic impl of CtSelect for [T; N] where T = u8.

Ideally we would use specialization to provide more specific impls of these traits for these types, but it’s unstable and unlikely to be stabilized soon.

Required Methods

fn bytes_ct_assign(&mut self, other: &Self, choice: Choice)

Conditionally assign other to self if choice is Choice::TRUE.

fn bytes_ct_select(&self, other: &Self, choice: Choice) -> Self

Select between self and other based on choice, returning a copy of the value.

Returns

self if choice is Choice::FALSE.

other if choice is Choice::TRUE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.