ctutils

Trait CtAssign

pub trait CtAssign<Rhs: ?Sized = Self> {
    // Required method
    fn ct_assign(&mut self, src: &Rhs, choice: Choice);
}

Constant-time conditional assignment: assign a given value to another based on a Choice.

This crate provides built-in implementations for the following types:

i8, i16, i32, i64, i128, isize

u8, u16, u32, u64, u128, usize

NonZeroI8, NonZeroI16, NonZeroI32, NonZeroI64, NonZeroI128, NonZeroI128

NonZeroU8, NonZeroU16, NonZeroU32, NonZeroU64, NonZeroU128,, NonZeroUsize

cmp::Ordering

Choice

[T] and [T; N] where T impls CtAssignSlice, which the previously mentioned types all do.

Required Methods

fn ct_assign(&mut self, src: &Rhs, choice: Choice)

Conditionally assign src to self if choice is Choice::TRUE.