ctutils

Trait CtAssignSlice

pub trait CtAssignSlice: CtAssign + Sized {
    // Provided method
    fn ct_assign_slice(dst: &mut [Self], src: &[Self], choice: Choice) { ... }
}

Implementing this trait enables use of the CtAssign trait for [T] where T is the Self type implementing the trait, via a blanket impl.

It needs to be a separate trait from CtAssign because we need to be able to impl CtAssign for [T] which is ?Sized.

Provided Methods

fn ct_assign_slice(dst: &mut [Self], src: &[Self], choice: Choice)

Conditionally assign src to dst if choice is Choice::TRUE, or leave it unchanged for Choice::FALSE.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.