ctutils::traits

Module ct_select

Macros

impl_ct_select_with_ct_assign 🔒

Macro to write impls of CtSelectUsingCtAssign.

Traits

CtSelect

Constant-time selection: choose between two values based on a given Choice.

CtSelectArray

Implementing this trait enables use of the CtSelect trait to construct [T; N] where T is the Self type implementing the trait, via a blanket impl.

CtSelectUsingCtAssign

Marker trait which enables a blanket impl of CtSelect for types which also impl Clone + CtAssign.