ctutils

Trait CtEqSlice

pub trait CtEqSlice: CtEq + Sized {
    // Provided methods
    fn ct_eq_slice(a: &[Self], b: &[Self]) -> Choice { ... }
    fn ct_ne_slice(a: &[Self], b: &[Self]) -> Choice { ... }
}

Implementing this trait enables use of the CtEq trait for [T] where T is the Self type implementing the trait, via a blanket impl.

It needs to be a separate trait from CtEq because we need to be able to impl CtEq for [T] which is ?Sized.

Provided Methods

fn ct_eq_slice(a: &[Self], b: &[Self]) -> Choice

Determine if a is equal to b in constant-time.

fn ct_ne_slice(a: &[Self], b: &[Self]) -> Choice

Determine if a is NOT equal to b in constant-time.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.