ctutils

Trait CtLookup

pub trait CtLookup<Idx> {
    type Output: CtAssign;

    // Required method
    fn ct_lookup(&self, index: Idx) -> CtOption<Self::Output>;
}

Constant-time lookup by index, similar to the Index trait, but returning an owned result in constant-time.

Required Associated Types

type Output: CtAssign

Output type returned by the lookup operation.

Required Methods

fn ct_lookup(&self, index: Idx) -> CtOption<Self::Output>

Attempt to retrieve the item at the given index, either returning it or the CtOption equivalent of None if the index was out-of-bounds.